backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
google-genai>=1.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
PyJWT>=2.8.0
pymupdf>=1.24.0
httpx>=0.27.0
trafilatura>=1.6.0
lightrag-hku>=0.1.0
numpy>=1.24.0
